# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Sorokchat - Український месенджер
  description: Sorokchat - Український месенджер
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'

paths:
  /authorization/register:
    post:
      operationId: /authrization/register
      summary: Реєстрація
      description: Реєстрація користувача у системі
      tags:
        - Авторизація
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUserPayload'
      responses:
        201:
          description: Успішна реєстрація
          headers:
            Location:
              required: true
              description: Посилання на себе
              example: /authorization/profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizedPayload'
        409:
          description: Користувач вже існує
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        400:
          description: Не коректні данні
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        403:
          description: Вже авторизований користувач
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /authorization/login:
    post:
      tags:
        - Авторизація
      summary: Вхід
      operationId: /authorization/login
      description: Вхід користувача у системі.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginPayload'
      responses:
        200:
          description: Успішний вхід
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizedPayload'
        404:
          description: Користувач не знайдений
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        400:
          description: Не коректні данні
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        403:
          description: Вже авторизований користувач
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /authorization/logout:
    delete:
      tags:
        - Авторизація
      summary: Вихід
      description: Вихід користувача із системи
      operationId: /authorization/logout
      security:
        - bearerAuthorization: [ ]
      responses:
        204:
          description: Успішний вихід
        401:
          description: Неавторизований користувач
          headers:
            WWW-AUTHENTICATE:
              required: true
              description: Спосіб авторизації
              schema:
                type: string
                example: Bearer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /authorization/profile:
    get:
      tags:
        - Авторизація
      summary: Профіль
      description: Отримання даних про себе
      operationId: /authorization/profile
      security:
        - bearerAuthorization: [ ]
      responses:
        200:
          description: Успішно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserPayload'
        401:
          description: Неавторизований користувач
          headers:
            WWW-AUTHENTICATE:
              required: true
              description: Спосіб авторизації
              schema:
                type: string
                example: Bearer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /authorization/refresh-tokens:
    put:
      tags:
        - Авторизація
      operationId: /authorization/refresh-tokens
      description: Оновлення токенів у системі.
      summary: Оновлення токенів.
      responses:
        200:
          description: Успішно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizedPayload'
        401:
          description: Неавторизований користувач.
          headers:
            WWW-AUTHENTICATE:
              required: true
              description: Спосіб авторизації.
              schema:
                type: string
                example: Form
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /users:
    patch:
      tags:
        - Користувачі
      operationId: users/patch
      summary: Оновлення себе.
      description: Оновлення даних користувача у системі.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserPayload'
      security:
        - bearerAuthorization: [ ]
      responses:
        200:
          description: Успішне оновлення
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserPayload'
        401:
          description: Неавторизований користувач.
          headers:
            WWW-AUTHENTICATE:
              required: true
              description: Спосіб авторизації.
              schema:
                type: string
                example: Form
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        409:
          description: Унікальні данні конфліктують з іншими користувачами.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        400:
          description: Некоректні данні.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    delete:
      tags:
        - Користувачі
      security:
        - bearerAuthorization: [ ]
      summary: Видалення.
      description: Видалення користувачів із системи.
      operationId: /users/delete
      responses:
        204:
          description: Успішне видалення
        401:
          description: Неавторизований користувач.
          headers:
            WWW-AUTHENTICATE:
              required: true
              description: Спосіб авторизації.
              schema:
                type: string
                example: Form
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /users/{term}:
    get:
      tags:
        - Користувачі
      summary: Пошук користувача.
      description: Пошук користувача у системі.
      operationId: /users/term
      security:
        - bearerAuthorization: [ ]
      parameters:
        - name: term
          in: path
          required: true
          schema:
            type: string
            description: Фільтр для пошуку. Необов'язкове співпадіння.
            example: "@sorokovsky-andrey || Сороковський"
      responses:
        200:
          description: Успішний пошук.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetUserPayload'
        401:
          description: Неавторизований користувач.
          headers:
            WWW-AUTHENTICATE:
              required: true
              description: Спосіб авторизації.
              schema:
                type: string
                example: Form
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        404:
          description: Користувач не знайдений.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /contacts/create/{userNickname}:
    post:
      tags:
        - Контакти
      summary: Створення контакту.
      description: Створення контакту у системі для 2-х користувачів.
      operationId: /contacts/create
      parameters:
        - in: path
          name: userNickname
          required: true
          schema:
            type: string
            description: Унікальне ім'я користувача. Співпадіння обов'язкове.
            example: sorokovsky-andrey
      security:
        - bearerAuthorization: [ ]
      responses:
        201:
          description: Успішне створення контакту.
          headers:
            Location:
              required: true
              description: Посилання на створений контакт.
              example: /contacts/1
        401:
          description: Неавторизований користувач.
          headers:
            WWW-AUTHENTICATE:
              required: true
              description: Спосіб авторизації.
              schema:
                type: string
                example: Form
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        404:
          description: Користувач не знайдений.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /contacts/by-user/{term}:
    get:
      tags:
        - Контакти
      operationId: /contacts/get-one
      summary: Контакт.
      description: Отримання одного контакту.
      security:
        - bearerAuthorization: [ ]
      parameters:
        - in: path
          name: term
          schema:
            type: string
            description: Данні для пошуку контакту по унікальному імені. Співпадіння необов'язкова.
            example: "@sor || Сороковський"
      responses:
        200:
          description: Успішне отримання.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetContactPayload'
        401:
          description: Неавторизований користувач.
          headers:
            WWW-AUTHENTICATE:
              required: true
              description: Спосіб авторизації.
              schema:
                type: string
                example: Form
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        404:
          description: Користувач не знайдений.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /contacts/my:
    get:
      summary: Контакти користувача.
      operationId: contacts/userId
      security:
        - bearerAuthorization: [ ]
      tags:
        - Контакти
      responses:
        200:
          description: Успішне отримання.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetContactPayload'
        401:
          description: Неавторизований користувач.
          headers:
            WWW-AUTHENTICATE:
              required: true
              description: Спосіб авторизації.
              schema:
                type: string
                example: Form
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        404:
          description: Користувач не знайдений.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
components:
  securitySchemes:
    bearerAuthorization:
      type: http
      scheme: bearer
  schemas:
    NewUserPayload:
      type: object
      description: Данні для створення користувача.
      required:
        - nickname
        - displayName
        - password
      properties:
        nickname:
          type: string
          description: Унікальне ім'я користувача.
          example: sorokovsky-andrey
        displayName:
          type: string
          description: Видиме ім'я користувача.
          example: Сороковський Андрій
        password:
          type: string
          description: Пароль користувача.
          example: <PASSWORD>
    LoginPayload:
      description: Данні для входу
      required:
        - nickname
        - password
      properties:
        nickname:
          type: string
          description: Унікальне ім'я користувача.
          example: sorokovsky-andrey
        password:
          type: string
          description: Пароль користувача.
          example: <PASSWORD>
    UpdateUserPayload:
      type: object
      description: Данні для оновлення користувача.
      properties:
        nickname:
          type: string
          description: Унікальне ім'я користувача.
          example: sorokovsky-andrey
        displayName:
          type: string
          description: Видиме ім'я користувача.
          example: Сороковський Андрій
        password:
          type: string
          description: Пароль користувача.
          example: <PASSWORD>
        phoneNumber:
          type: string
          description: Номер телефону користувача.
          example: +XXXXXXXXXXXXX
        email:
          type: string
          description: Електронна адреса користувача.
          format: email
          example: support@example.com
    GetUserPayload:
      type: object
      description: Данні для отримання користувача.
      required:
        - id
        - createdAt
        - updatedAt
        - nickname
        - displayName
        - password
      properties:
        id:
          type: integer
          description: Ідентифікатор користувача
          example: 1
        createdAt:
          type: string
          format: timestamp
          description: Дата створення користувача.
          example: 14.01.2026T13:25
        updatedAt:
          type: string
          description: Дата оновлення користувача.
          format: timestamp
          example: 14.01.2026T13:25
        nickname:
          type: string
          description: Унікальне ім'я користувача.
          example: sorokovsky-andrey
        displayName:
          type: string
          description: Видиме ім'я користувача.
          example: Сороковський Андрій
        password:
          type: string
          description: Пароль користувача.
          example: <PASSWORD>
        phoneNumber:
          type: string
          description: Номер телефону користувача.
          example: +XXXXXXXXXXXXX
        email:
          type: string
          description: Електронна адреса користувача.
          format: email
          example: support@example.com
    GetContactPayload:
      type: object
      description: Данні для отримання користувача.
      required:
        - id
        - createdAt
        - updatedAt
        - firstUser
        - secondUser
      properties:
        id:
          type: integer
          description: Ідентифікатор контакту.
          example: 1
        createdAt:
          type: string
          format: timestamp
          description: Дата створення контакту.
          example: 14.01.2026T13:25
        updatedAt:
          type: string
          description: Дата оновлення контакту.
          format: timestamp
          example: 14.01.2026T13:25
        firstUser:
          $ref: '#/components/schemas/GetUserPayload'
        secondUser:
          $ref: '#/components/schemas/GetUserPayload'
    ProblemDetail:
      type: object
      description: Стандартний формат відповіді про помилку (RFC 9457).
      properties:
        type:
          type: string
          format: uri
          description: URI-посилання, яке ідентифікує тип проблеми. Зазвичай вказує на документацію.
          example: https://example.com/probs/out-of-stock
        title:
          type: string
          description: Короткий, зрозумілий для людини опис типу проблеми.
          example: Помилка реєстрації
        status:
          type: integer
          description: HTTP-код статусу, згенерований сервером.
          example: 400
        detail:
          type: string
          description: Детальне пояснення конкретної помилки в поточному контексті.
          example: Не коректна електронна адреса.
        instance:
          type: string
          format: uri
          description: URI-посилання на конкретне виникнення проблеми (наприклад, шлях запиту).
          example: /authorization/register
      additionalProperties: true
    AuthorizedPayload:
      description: Данні після авторизації.
      required:
        - accessToken
      properties:
        accessToken:
          type: string
          description: Токен доступу.
          example: <JWS TOKEN>